+++ odoo/local-src/specific_sale/models/__init__.py
@@ -8,3 +8,4 @@ from . import sale_order
 from . import mail_compose_message
 from . import mrp_bom
 from . import project_task
+from . import procurement_order